Pantry-Detection-main/Web Application/public/cloud/annotated_image/requirements.txt
ultralytics
opencv-python
google-cloud-storage